+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/buttons/action/HideButtonsPatch.kt
@@ -28,7 +28,8 @@ val hideButtonsPatch = resourcePatch(
             "19.43.41",
             "19.47.53",
             "20.07.39",
-        ),
+            "20.12.46",
+        )
     )
 
     execute {
@@ -45,10 +46,11 @@ val hideButtonsPatch = resourcePatch(
                     SwitchPreference("revanced_hide_remix_button"),
                     SwitchPreference("revanced_hide_download_button"),
                     SwitchPreference("revanced_hide_thanks_button"),
+                    SwitchPreference("revanced_hide_ask_button"),
                     SwitchPreference("revanced_hide_clip_button"),
                     SwitchPreference("revanced_hide_playlist_button"),
-                ),
-            ),
+                )
+            )
         )
 
         addLithoFilter("Lapp/revanced/extension/youtube/patches/components/ButtonsFilter;")