+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/searchbar/WideSearchbarPatch.kt
@@ -1,30 +1,67 @@
 package app.revanced.patches.youtube.layout.searchbar
 
-import app.revanced.patcher.Fingerprint
-import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
+import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
-import app.revanced.patcher.patch.BytecodePatchContext
 import app.revanced.patcher.patch.bytecodePatch
-import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod
+import app.revanced.patcher.patch.resourcePatch
 import app.revanced.patches.all.misc.resources.addResources
 import app.revanced.patches.all.misc.resources.addResourcesPatch
+import app.revanced.patches.shared.misc.mapping.get
+import app.revanced.patches.shared.misc.mapping.resourceMappingPatch
+import app.revanced.patches.shared.misc.mapping.resourceMappings
 import app.revanced.patches.shared.misc.settings.preference.SwitchPreference
 import app.revanced.patches.youtube.misc.extension.sharedExtensionPatch
 import app.revanced.patches.youtube.misc.settings.PreferenceScreen
 import app.revanced.patches.youtube.misc.settings.settingsPatch
+import app.revanced.util.addInstructionsAtControlFlowLabel
+import app.revanced.util.findInstructionIndicesReversedOrThrow
+import app.revanced.util.getReference
+import app.revanced.util.indexOfFirstInstructionOrThrow
+import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
+import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 
 private const val EXTENSION_CLASS_DESCRIPTOR =
     "Lapp/revanced/extension/youtube/patches/WideSearchbarPatch;"
 
+internal var ytWordmarkHeaderId = -1L
+    private set
+internal var ytPremiumWordmarkHeaderId = -1L
+    private set
+internal var actionBarRingoId = -1L
+    private set
+
+private val wideSearchbarResourcePatch = resourcePatch {
+    dependsOn(resourceMappingPatch)
+
+    execute {
+        ytWordmarkHeaderId = resourceMappings[
+            "attr",
+            "ytWordmarkHeader",
+        ]
+
+        ytPremiumWordmarkHeaderId = resourceMappings[
+            "attr",
+            "ytPremiumWordmarkHeader",
+        ]
+
+        actionBarRingoId = resourceMappings[
+            "layout",
+            "action_bar_ringo",
+        ]
+    }
+}
+
 val wideSearchbarPatch = bytecodePatch(
     name = "Wide search bar",
-    description = "Adds an option to replace the search icon with a wide search bar. This will hide the YouTube logo when active.",
+    description = "Adds an option to replace the search icon with a wide search bar. " +
+            "This will hide the YouTube logo when active.",
 ) {
     dependsOn(
         sharedExtensionPatch,
         settingsPatch,
         addResourcesPatch,
+        wideSearchbarResourcePatch,
     )
 
     compatibleWith(
@@ -35,7 +72,8 @@ val wideSearchbarPatch = bytecodePatch(
             "19.43.41",
             "19.47.53",
             "20.07.39",
-        ),
+            "20.12.46",
+        )
     )
 
     execute {
@@ -45,37 +83,45 @@ val wideSearchbarPatch = bytecodePatch(
             SwitchPreference("revanced_wide_searchbar"),
         )
 
-        /**
-         * Navigate a fingerprints method at a given index mutably.
-         *
-         * @param index The index to navigate to.
-         * @param from The fingerprint to navigate the method on.
-         * @return The [MutableMethod] which was navigated on.
-         */
-        fun BytecodePatchContext.walkMutable(index: Int, from: Fingerprint) =
-            navigate(from.originalMethod).to(index).stop()
+        setWordmarkHeaderFingerprint.let {
+            // Navigate to the method that checks if the YT logo is shown beside the search bar.
+            val shouldShowLogoMethod = with(it.originalMethod) {
+                val invokeStaticIndex = indexOfFirstInstructionOrThrow {
+                    opcode == Opcode.INVOKE_STATIC &&
+                            getReference<MethodReference>()?.returnType == "Z"
+                }
+                navigate(this).to(invokeStaticIndex).stop()
+            }
 
-        /**
-         * Injects instructions required for certain methods.
-         */
-        fun MutableMethod.injectSearchBarHook() {
-            val insertIndex = implementation!!.instructions.size - 1
-            val insertRegister = getInstruction<OneRegisterInstruction>(insertIndex).registerA
+            shouldShowLogoMethod.apply {
+                findInstructionIndicesReversedOrThrow(Opcode.RETURN).forEach { index ->
+                    val register = getInstruction<OneRegisterInstruction>(index).registerA
 
-            addInstructions(
-                insertIndex,
-                """
-                invoke-static {v$insertRegister}, $EXTENSION_CLASS_DESCRIPTOR->enableWideSearchbar(Z)Z
-                move-result v$insertRegister
-            """,
-            )
+                    addInstructionsAtControlFlowLabel(
+                        index,
+                        """
+                            invoke-static { v$register }, $EXTENSION_CLASS_DESCRIPTOR->enableWideSearchbar(Z)Z
+                            move-result v$register
+                        """
+                    )
+                }
+            }
         }
 
-        mapOf(
-            setWordmarkHeaderFingerprint to 1,
-            createSearchSuggestionsFingerprint to createSearchSuggestionsFingerprint.patternMatch!!.startIndex,
-        ).forEach { (fingerprint, callIndex) ->
-            walkMutable(callIndex, fingerprint).injectSearchBarHook()
+        // Fix missing left padding when using wide searchbar.
+        wideSearchbarLayoutFingerprint.method.apply {
+            findInstructionIndicesReversedOrThrow {
+                val reference = getReference<MethodReference>()
+                reference?.definingClass == "Landroid/view/LayoutInflater;"
+                        && reference.name == "inflate"
+            }.forEach { inflateIndex ->
+                val register = getInstruction<OneRegisterInstruction>(inflateIndex + 1).registerA
+
+                addInstruction(
+                    inflateIndex + 2,
+                    "invoke-static { v$register }, $EXTENSION_CLASS_DESCRIPTOR->setActionBar(Landroid/view/View;)V"
+                )
+            }
         }
     }
 }